style: add style to `MyPage`
